+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Change Log
 
-All notable changes to the "daily-bullet-notes" extension will be documented in this file.
+## 0.0.3 - 2024-12-5
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+* Fixed bug where `[x]` or `[-]` sub-tasks were not removed on next day
+* `[ ]` is now highlighted just like `[]`
+* Added [New DBM File](command:daily-bullet-notes.newFile) command which will create a new file with a template to get new users started
+* Readme updates to get new users started
 
-## [Unreleased]
+## 0.0.2 - 2024-12-3
 
-- Initial release
+* Updated readme with extension overview
+* Progress/blocked/etc markers are removed from task boxes when copied to today so your new day starts with a clean slate
+
+## 0.0.1 - 2024-12-1
+
+* Initial version - not yet complete, many bugs